New approach in makers
New approach adopted for info maker and other container
makers. The idea is instead of specifying different
types for each function, we shall overload a function
with pattern matching using data constructors of given
types.
This should significantly diminish the code needed
for typeclasses of unitmodel and containermodel
and of course containerdata model