Complete change in the overall layout
Complete change in the overall layout. Now
typeclasses are under DataUtils.hs
And models have their own namespace.
Models reuse typeclasses as required.
The rendering is done through view.
So far only xml is in the way of implementation.
Unit and Container should be rendered as an Element.
Unit and Container should also be rendered as a node.
Once the xml rendering is done, I should concentrate
on io.

Model types should have a corresponding specification.
The specification file should have a schema. Maybe in json
or yaml.
The spefication should contain the attributes that are available
for each model type.