Another huge refactoring on the underlaying structure
Another huge refactoring on the underlaying structure of the project
Complete adoption of mvc structure is on its way.
Typeclasses for controller is on its way as well.
Here is a global view on the changes:

Anything that takes in model as input and outputs a native representation
like a string or text is considered as a view utility.
Thus we put it into viewutils. Anything that transforms a native
representation to model is considered as model utility.
Thuw we put it into modelutils. Anything that modifies models
are considered as a control utility thus it has been put to
controlutils.hs

ModelData type is introduced in order to unify control
typeclass. It is pure data with no info. It can be constructed
from a unitdata or container data. Unit model continues to
use unitdata as its model data, container model also continues
to use container data as its model data, but ReplaceField
typeclass on controller specifies modeldata as an argument to
replaceData. Each instance should implement 2 versions of
the replaceData function, one with the appropriate constructor
the other with the inappropriate constructor.
MatchModel typeclass is not yet implemented