Refactoring continues
Implementing setters with transformer was a good idea.
ModelInfo should be serialized to tuple like structure.
Map look up is by its nature difficult to do it in completely
pure state. Because what if key does not exist. It is better
to rely on order or position of the element.
ModelInfo should be set into three elements, String, String,
Map String String, and not just a Map String String.