Refactoring continues, program compiles
Refactoring continues but I should focus on tests now.
I shall not change anything until I have written some
tests, in order not to break compilation. And it is
getting difficult to imagine the functionality.

Here are few new ideas to implement:

- ModelId, ModelType, ModelAttr can have impure functions for
setting and transforming.
- UnitData, thus model data as well can have an impure implementation
 for setting and transforming functions.
- All the rest of computations should stay within the pure ground.

The challenge is to implement a pure setting function for container data.
since it uses unitmodel and container model as value in its data
constructor.
But basically if it has something that thing is a container data, if
that thing has just content, content is unit data and that thing is
a unit model, if it has other things then that thing is a container
model.
This maps to xml as the following, if the contents of an element is
empty or decomposes into content node, then that node is a unitmodel
if contents of an element decomposes into element node, then
that node is a container model and its contents are container data.